requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
pdfplumber
pyyaml
openpyxl
pytesseract
Pillow
jinja2